Generalize TANE and PFDTANE algorithms
