+++ src/core/algorithms/fd/mining_algorithms.h
@@ -8,6 +8,6 @@
 #include "algorithms/fd/fdep/fdep.h"
 #include "algorithms/fd/fun/fun.h"
 #include "algorithms/fd/hyfd/hyfd.h"
-#include "algorithms/fd/pfdtane/pfdtane.h"
 #include "algorithms/fd/pyro/pyro.h"
+#include "algorithms/fd/tane/pfdtane.h"
 #include "algorithms/fd/tane/tane.h"
