+++ src/core/algorithms/fd/tane/tane.h
@@ -1,47 +1,21 @@
 #pragma once
 
-#include <string>
-
-#include "algorithms/fd/pli_based_fd_algorithm.h"
 #include "config/error/type.h"
+#include "model/table/column_data.h"
 #include "model/table/position_list_index.h"
-#include "model/table/relation_data.h"
+#include "tane_common.h"
 
 namespace algos {
 
-class Tane : public PliBasedFDAlgorithm {
+class Tane : public tane::TaneCommon {
 private:
-    void RegisterOptions();
-    void MakeExecuteOptsAvailableFDInternal() final;
-
-    void ResetStateFd() final;
-    unsigned long long ExecuteInternal() final;
+    void MakeExecuteOptsAvailableFDInternal() override final;
+    config::ErrorType CalculateZeroAryFdError(ColumnData const* rhs) override;
+    config::ErrorType CalculateFdError(model::PositionListIndex const* lhs_pli,
+                                       model::PositionListIndex const* joint_pli) override;
 
 public:
-    config::ErrorType max_fd_error_;
-    config::ErrorType max_ucc_error_;
-
-    int count_of_fd_ = 0;
-    int count_of_ucc_ = 0;
-    long apriori_millis_ = 0;
-
     Tane(std::optional<ColumnLayoutRelationDataManager> relation_manager = std::nullopt);
-
-    static double CalculateZeroAryFdError(ColumnData const* rhs,
-                                          ColumnLayoutRelationData const* relation_data);
-    static double CalculateFdError(model::PositionListIndex const* lhs_pli,
-                                   model::PositionListIndex const* joint_pli,
-                                   ColumnLayoutRelationData const* relation_data);
-    static double CalculateUccError(model::PositionListIndex const* pli,
-                                    ColumnLayoutRelationData const* relation_data);
-
-    // static double round(double error) { return ((int)(error * 32768) + 1)/ 32768.0; }
-
-    void RegisterAndCountFd(Vertical const& lhs, Column const* rhs, double error,
-                            RelationalSchema const* schema);
-    // void RegisterFd(Vertical const* lhs, Column const* rhs, double error, RelationalSchema const*
-    // schema);
-    void RegisterUcc(Vertical const& key, double error, RelationalSchema const* schema);
 };
 
 }  // namespace algos
